categories sql 모델 생성
